edit group commit

+++ regression-test/suites/datatype_p0/ip/test_ip_implicit_cast.groovy
@@ -17,6 +17,9 @@
 // under the License.
 
 suite("test_ip_implicit_cast") {
+    if (isGroupCommitMode()) {
+        return
+    }
     def tableName = "test_ip_implicit_cast"
     sql """ DROP TABLE IF EXISTS ${tableName} """
